Implement chainable async array transformers
This is a proof-of-concept implementation of the new chainable and
thenable asynchronous array transformation mechanism.

Changes summary:

 * Add new `AsyncArray` class which is an abstraction of an
   asynchronously retrieved array that can be transformed or iterated
   with asynchronous functions.
 * Add new `metasync.for(arrayOrPromise)` function to instantiate an
   `AsyncArray`.
 * Move array-related metasync functions to a separate module to avoid
   cyclic dependency.
 * Add a simple test.

TODO list:

 * Investigate the performance cost of such solution, maybe it can be a
   lightweight thenable without the need for full-blown promises, or
   there may be other possible optimizations.
 * Maybe implement a similar mechanism for any iterable objects, not
   just arrays.
 * Maybe support functions returning promises as callbacks of
   `AsyncArray` methods.

Refs: https://github.com/metarhia/MetaSync/issues/17#issuecomment-280790478
Refs: https://github.com/metarhia/MetaSync/issues/37
